Refactor number format to be testable
