+++ apps/patient/src/utils/formatters.test.ts
@@ -1,0 +1,5 @@
+import { formatPrice } from './formatters';
+
+test('formatPrice', () => {
+  expect(formatPrice(100)).toEqual('$100.00');
+});
