+++ apps/patient/src/utils/formatters.ts
@@ -1,0 +1,3 @@
+export const formatPrice = (price: number) => {
+  return Number.isInteger(price) ? Math.round(price) : price.toFixed(2);
+};
